some new models

+++ Models/Refrigerator.cs
@@ -9,7 +9,7 @@ namespace HomeElectronics.Models
     public class Refrigerator : NoFrostTech
     {
         public override string Name { get; set; } = "Refrigerator";
-        public override int PowerConsumption { get; set; }
+        public override double PowerConsumption { get; set; }
         public override string CompanyName { get; set; }
     }
 }
